requirements.txt
Django==2.0.4
djangorestframework==3.8.2
mysqlclient==1.3.12
Pillow==5.1.0